added confirmation to DeleteUserButton

+++ client/src/components/UI/DeleteUserButton.jsx
@@ -5,19 +5,26 @@ import { DELETE_USER } from '../../utils/mutations';
 import Auth from '../../utils/auth';
 
 function DeleteUserButton() {
-    const { loading, data } = useQuery(QUERY_ME);
+  const { loading, data } = useQuery(QUERY_ME);
   const [deleteUser] = useMutation(DELETE_USER);
 
   const handleDeleteUser = async () => {
     try {
       if (data && data.me) {
-        const { data: deleteUserData } = await deleteUser({
-          variables: { id: data.me._id },
-        });
-  
-        // Handle success or error
-        console.log('User deleted:', deleteUserData.deleteUser);
-        Auth.logout();
+        // Create a custom confirmation box
+        const confirmationPhrase = window.prompt('To confirm account deletion, please type "DELETE MY ACCOUNT" (case-sensitive):');
+
+        if (confirmationPhrase === 'DELETE MY ACCOUNT') {
+          const { data: deleteUserData } = await deleteUser({
+            variables: { id: data.me._id },
+          });
+
+          // Handle success or error
+          console.log('User deleted:', deleteUserData.deleteUser);
+          Auth.logout();
+        } else {
+          alert('Confirmation phrase did not match. Account deletion canceled.');
+        }
       }
     } catch (error) {
       console.error('Error deleting user:', error);
